move shared/test/middleware to dashboard/test/lib
