+++ dashboard/test/lib/middleware/files_api_test_base.rb
@@ -1,9 +1,9 @@
 require_relative 'middleware_test_helper' # Must be required first to establish load paths
 require 'mocha/mini_test'
-require_relative '../../middleware/files_api'
-require_relative '../../middleware/channels_api'
+require_relative '../../../lib/middleware/files_api'
+require_relative '../../../lib/middleware/channels_api'
 require 'cdo/aws/s3'
-require_relative '../spy_newrelic_agent'
+require_relative '../../../../shared/test/spy_newrelic_agent'
 
 #
 # Base class for tests against the FilesApi (which include SourcesTest,
