+++ dashboard/test/lib/middleware/helpers/test_animation_bucket.rb
