+++ dashboard/test/lib/middleware/helpers/test_asset_bucket.rb
