+++ dashboard/test/lib/middleware/test_animations.rb
