+++ dashboard/test/lib/middleware/test_files.rb
