+++ dashboard/test/lib/middleware/test_libraries.rb
